backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.0
python-dotenv==1.0.1

